add page views events
